requirements.txt
fastapi
mongoengine
SQLAlchemy
# mysql-connector
mysql-connector-python
Flask
Flask-JWT-Extended
python-dotenv
PyMySQL[rsa]
requests